Move model repositories to separate file.
Prep for later switch of models to declarative style.
